Consolidate stock upload logic
